+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [unreleased]
+
+### pack
+
+* Added `--react-compiler` option to build using `babel-plugin-react-compiler`
+
 ## 7.3.2 (June 23, 2026)
 
 ### create, template